% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dwellings.R
\name{equal_dwellings}
\alias{equal_dwellings}
\title{Heatmaps of Sites with Equal Dwellings}
\usage{
equal_dwellings(nrow, ncol, dwellings, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{dwellings}{Number of dwellings}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and a printing of the number of combinations in the console.
}
\description{
Function for creating heatmaps of sites that have an equal number of dwellings.
}
\examples{
equal_dwellings(nrow = 3, ncol = 2, dwellings = 4, filename = "equal_dwellings_nrow3_ncol2_dwellings4.pdf")
}
