% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_far.R
\name{equal_far}
\alias{equal_far}
\title{Heatmaps of Sites with Equal Floor Area Ratio (FAR)}
\usage{
equal_far(nrow, ncol, far, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{far}{Floor Area Ratio as a fraction}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and the number of combinations.
}
\description{
Function for creating heatmaps of sites that have an equal measure of FAR.
}
\examples{
equal_far(nrow = 1, ncol = 3, far = 6/3, filename = "equal_far_nrow1_ncol3_far2.pdf")
}
