% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_gsi.R
\name{equal_gsi}
\alias{equal_gsi}
\title{Heatmaps of Sites with Equal Ground Space Index (GSI)}
\usage{
equal_gsi(nrow, ncol, gsi, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{gsi}{GSI as a fraction between zero and one}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and a printing of the number of combinations in the console.
}
\description{
Function for creating heatmaps of sites that have an equal GSI.
}
\examples{
equal_gsi(nrow = 3, ncol = 3, gsi = 2/3, filename = "equal_gsi_nrow3_ncol3_gsi2div3.pdf")
}
