% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_rcr.R
\name{equal_rcr}
\alias{equal_rcr}
\title{Heatmaps of Sites with Equal Roof Coverage Ratio (RCR)}
\usage{
equal_rcr(nrow, ncol, rcr)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{rcr}{RCR as a fraction between zero and one}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for creating heatmaps of sites that have an equal RCR.
}
\examples{
equal_rcr(nrow = 3, ncol = 3, rcr = 7/9)
}
