% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_units.R
\name{equal_units}
\alias{equal_units}
\title{Heatmaps of Floorplans with an Equal Number of Units}
\usage{
equal_units(nrow, ncol, units)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{units}{Number of units}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for comparing the adaptability of floorplans by creating heatmaps with an equal number of units.
}
\examples{
equal_units(nrow = 4, ncol = 3, units = 4)
}
