% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_far.R
\name{max_far}
\alias{max_far}
\title{Heatmaps of Building Combinations with Measure of Max FAR}
\usage{
max_far(nrow, ncol, max_far, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{max_far}{Maximum FAR as a fraction}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and a printing of the number of combinations in the console.
}
\description{
Function for specifying maximum FAR and creating heatmaps that show all combinations.
}
\examples{
max_far(nrow = 2, ncol = 2, max_far = 6/4, filename = "max_far_nrow2_ncol2_maxfar6div4.pdf")
}
