% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_storeys.R
\name{max_storeys}
\alias{max_storeys}
\title{Heatmaps of Building Combinations with Measure of Max Storeys}
\usage{
max_storeys(nrow, ncol, max_storey, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{max_storey}{Maximum storey}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and a printing of the number of combinations in the console.
}
\description{
Function for specifying maximum storeys and creating heatmaps that show all combinations.
}
\examples{
max_storeys(nrow = 1, ncol = 3, max_storey = 3, filename = "max_storeys_nrow1_ncol3_maxstorey3.pdf")
}
