% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_far.R
\name{min_max_far}
\alias{min_max_far}
\title{Heatmaps of Building Combinations with Measure of Min and Max FAR}
\usage{
min_max_far(nrow, ncol, min_far, max_far)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{min_far}{Minimum FAR as a fraction}

\item{max_far}{Maximum FAR as a fraction}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for specifying minimum and maximum FAR and creating heatmaps that show all combinations.
}
\examples{
min_max_far(nrow = 2, ncol = 2, min_far = 3/4, max_far = 6/4)
}
