% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_storeys.R
\name{min_max_storeys}
\alias{min_max_storeys}
\title{Heatmaps of Building Combinations with Measure of Min and Max Storey}
\usage{
min_max_storeys(nrow, ncol, min_storey, max_storey, filename)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{min_storey}{Minimum storey}

\item{max_storey}{Maximum storey}

\item{filename}{Filename for PDF}
}
\value{
A PDF with multiple heatmaps and the number of combinations.
}
\description{
Function for specifying minimum and maximum storey and creating heatmaps that show all combinations.
}
\examples{
min_max_storeys(nrow = 1, ncol = 3, min_storey = 3, max_storey = 6, filename = "min_max_storeys_nrow1_ncol3_min3_max6.pdf")
}
