% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dwellings.R
\name{equal_dwellings}
\alias{equal_dwellings}
\title{Heatmaps of Sites with Equal Dwellings}
\usage{
equal_dwellings(nrow, ncol, dwellings)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{dwellings}{Number of dwellings}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for creating heatmaps of sites that have an equal number of dwellings.
}
\examples{
equal_dwellings(nrow = 3, ncol = 2, dwellings = 4)
}
