% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_far.R
\name{max_far}
\alias{max_far}
\title{Heatmaps of Building Combinations with Measure of Max FAR}
\usage{
max_far(nrow, ncol, max_far)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{max_far}{Maximum FAR as a fraction}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for specifying maximum FAR and creating heatmaps that show all combinations.
}
\examples{
max_far(nrow = 2, ncol = 2, max_far = 6/4)
}
