% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_storeys.R
\name{min_max_storeys}
\alias{min_max_storeys}
\title{Heatmaps of Building Combinations with Measure of Min and Max Storey}
\usage{
min_max_storeys(nrow, ncol, min_storey, max_storey)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{min_storey}{Minimum storey}

\item{max_storey}{Maximum storey}
}
\value{
Multiple heatmaps and a print-out of the number of combinations in the console.
}
\description{
Function for specifying minimum and maximum storey and creating heatmaps that show all combinations.
}
\examples{
min_max_storeys(nrow = 1, ncol = 3, min_storey = 3, max_storey = 6)
}
